rename: z index 파일명 컨벤션 수정
